split skill_toolset
