+++ core/src/common.ts
@@ -226,12 +226,10 @@ export {GoogleLLMVariant} from './utils/variant_utils.js';
 export {version} from './version.js';
 
 export type {Frontmatter, Resources, Script, Skill} from './skills/skill.js';
-export {
-  ListSkillsTool,
-  LoadSkillResourceTool,
-  LoadSkillTool,
-  SkillToolset,
-} from './tools/skill_toolset.js';
+export {ListSkillsTool} from './tools/skill/list_skills_tool.js';
+export {LoadSkillResourceTool} from './tools/skill/load_skill_resource_tool.js';
+export {LoadSkillTool} from './tools/skill/load_skill_tool.js';
+export {SkillToolset} from './tools/skill/skill_toolset.js';
 
 export * from './artifacts/base_artifact_service.js';
 export * from './memory/base_memory_service.js';
